Remove the ObjectMetadata concern, since it's only used in tests now
